v3_agentic_rag/v3_submission/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-google-genai==2.0.8
langchain-chroma==0.2.0
langgraph==0.2.60
chromadb==0.6.0
tavily-python==0.5.0
sentence-transformers==3.3.1
rank_bm25==0.2.2
streamlit==1.41.1
python-dotenv==1.0.1
pypdf==5.1.0
